requirements.txt
flask==3.0.3
pymupdf==1.24.5
requests==2.32.3
python-dotenv==1.0.1
